requirements.txt
customtkinter
python-docx
google-genai
python-dotenv
openai
anthropic
pypandoc
